pyproject.toml
[tool.poetry]
name = "drf-homework"
version = "0.1.0"
description = ""
authors = ["Valera K <vk1337atliberty@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.5.3"


[tool.poetry.group.dev.dependencies]
ipython = "^8.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
